rearrange and only handle PRICE_UPDATED
